Refactored Serializers into clearer folder structure
