+++ src/main/java/ldbc/snb/datagen/serializer/snb/csv/mergeforeign/CSVMergeForeignPersonActivitySerializer.java
@@ -33,7 +33,7 @@
  You should have received a copy of the GNU General Public License
  along with this program; if not, write to the Free Software
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.*/
-package ldbc.snb.datagen.serializer.snb.interactive;
+package ldbc.snb.datagen.serializer.snb.csv.mergeforeign;
 
 import ldbc.snb.datagen.dictionary.Dictionaries;
 import ldbc.snb.datagen.objects.*;
@@ -45,31 +45,23 @@ import java.io.IOException;
 import java.util.ArrayList;
 
 /**
- * @author aprat
+ * Created by aprat on 17/02/15.
  */
-public class CSVPersonActivitySerializer extends PersonActivitySerializer {
+public class CSVMergeForeignPersonActivitySerializer extends PersonActivitySerializer {
     private HDFSCSVWriter[] writers;
     private ArrayList<String> arguments;
     private String empty = "";
 
     private enum FileNames {
         FORUM("forum"),
-        FORUM_CONTAINEROF_POST("forum_containerOf_post"),
         FORUM_HASMEMBER_PERSON("forum_hasMember_person"),
-        FORUM_HASMODERATOR_PERSON("forum_hasModerator_person"),
         FORUM_HASTAG_TAG("forum_hasTag_tag"),
         PERSON_LIKES_POST("person_likes_post"),
         PERSON_LIKES_COMMENT("person_likes_comment"),
         POST("post"),
-        POST_HASCREATOR_PERSON("post_hasCreator_person"),
         POST_HASTAG_TAG("post_hasTag_tag"),
-        POST_ISLOCATEDIN_PLACE("post_isLocatedIn_place"),
         COMMENT("comment"),
-        COMMENT_HASCREATOR_PERSON("comment_hasCreator_person"),
-        COMMENT_HASTAG_TAG("comment_hasTag_tag"),
-        COMMENT_ISLOCATEDIN_PLACE("comment_isLocatedIn_place"),
-        COMMENT_REPLYOF_POST("comment_replyOf_post"),
-        COMMENT_REPLYOF_COMMENT("comment_replyOf_comment");
+        COMMENT_HASTAG_TAG("comment_hasTag_tag");
 
         private final String name;
 
@@ -97,23 +89,14 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
         arguments.add("id");
         arguments.add("title");
         arguments.add("creationDate");
+        arguments.add("moderator");
         writers[FileNames.FORUM.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
-        arguments.add("Forum.id");
-        arguments.add("Post.id");
-        writers[FileNames.FORUM_CONTAINEROF_POST.ordinal()].writeHeader(arguments);
         arguments.clear();
 
         arguments.add("Forum.id");
         arguments.add("Person.id");
         arguments.add("joinDate");
         writers[FileNames.FORUM_HASMEMBER_PERSON.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
-        arguments.add("Forum.id");
-        arguments.add("Person.id");
-        writers[FileNames.FORUM_HASMODERATOR_PERSON.ordinal()].writeHeader(arguments);
         arguments.clear();
 
         arguments.add("Forum.id");
@@ -141,22 +124,15 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
         arguments.add("language");
         arguments.add("content");
         arguments.add("length");
+        arguments.add("creator");
+        arguments.add("Forum.id");
+        arguments.add("place");
         writers[FileNames.POST.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
-        arguments.add("Post.id");
-        arguments.add("Person.id");
-        writers[FileNames.POST_HASCREATOR_PERSON.ordinal()].writeHeader(arguments);
         arguments.clear();
 
         arguments.add("Post.id");
         arguments.add("Tag.id");
         writers[FileNames.POST_HASTAG_TAG.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
-        arguments.add("Post.id");
-        arguments.add("Place.id");
-        writers[FileNames.POST_ISLOCATEDIN_PLACE.ordinal()].writeHeader(arguments);
         arguments.clear();
 
         arguments.add("id");
@@ -165,34 +141,17 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
         arguments.add("browserUsed");
         arguments.add("content");
         arguments.add("length");
+        arguments.add("creator");
+        arguments.add("place");
+        arguments.add("replyOfPost");
+        arguments.add("replyOfComment");
         writers[FileNames.COMMENT.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
-        arguments.add("Comment.id");
-        arguments.add("Person.id");
-        writers[FileNames.COMMENT_HASCREATOR_PERSON.ordinal()].writeHeader(arguments);
         arguments.clear();
 
         arguments.add("Comment.id");
         arguments.add("Tag.id");
         writers[FileNames.COMMENT_HASTAG_TAG.ordinal()].writeHeader(arguments);
         arguments.clear();
-
-        arguments.add("Comment.id");
-        arguments.add("Place.id");
-        writers[FileNames.COMMENT_ISLOCATEDIN_PLACE.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
-        arguments.add("Comment.id");
-        arguments.add("Post.id");
-        writers[FileNames.COMMENT_REPLYOF_POST.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
-        arguments.add("Comment.id");
-        arguments.add("Comment.id");
-        writers[FileNames.COMMENT_REPLYOF_COMMENT.ordinal()].writeHeader(arguments);
-        arguments.clear();
-
     }
 
     @Override
@@ -210,12 +169,8 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
         arguments.add(Long.toString(forum.id()));
         arguments.add(forum.title());
         arguments.add(dateString);
-        writers[FileNames.FORUM.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-        arguments.add(Long.toString(forum.id()));
         arguments.add(Long.toString(forum.moderator().accountId()));
-        writers[FileNames.FORUM_HASMODERATOR_PERSON.ordinal()].writeEntry(arguments);
+        writers[FileNames.FORUM.ordinal()].writeEntry(arguments);
         arguments.clear();
 
         for (Integer i : forum.tags()) {
@@ -237,23 +192,10 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
         arguments.add(Dictionaries.languages.getLanguageName(post.language()));
         arguments.add(post.content());
         arguments.add(Integer.toString(post.content().length()));
-        writers[FileNames.POST.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-        arguments.add(Long.toString(post.messageId()));
-        arguments.add(Integer.toString(post.countryId()));
-        writers[FileNames.POST_ISLOCATEDIN_PLACE.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-        arguments.add(Long.toString(post.messageId()));
         arguments.add(Long.toString(post.author().accountId()));
-        writers[FileNames.POST_HASCREATOR_PERSON.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-
         arguments.add(Long.toString(post.forumId()));
-        arguments.add(Long.toString(post.messageId()));
-        writers[FileNames.FORUM_CONTAINEROF_POST.ordinal()].writeEntry(arguments);
+        arguments.add(Integer.toString(post.countryId()));
+        writers[FileNames.POST.ordinal()].writeEntry(arguments);
         arguments.clear();
 
         for (Integer t : post.tags()) {
@@ -271,28 +213,16 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
         arguments.add(Dictionaries.browsers.getName(comment.browserId()));
         arguments.add(comment.content());
         arguments.add(Integer.toString(comment.content().length()));
-        writers[FileNames.COMMENT.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-        if (comment.replyOf() == comment.postId()) {
-            arguments.add(Long.toString(comment.messageId()));
-            arguments.add(Long.toString(comment.postId()));
-            writers[FileNames.COMMENT_REPLYOF_POST.ordinal()].writeEntry(arguments);
-            arguments.clear();
-        } else {
-            arguments.add(Long.toString(comment.messageId()));
-            arguments.add(Long.toString(comment.replyOf()));
-            writers[FileNames.COMMENT_REPLYOF_COMMENT.ordinal()].writeEntry(arguments);
-            arguments.clear();
-        }
-        arguments.add(Long.toString(comment.messageId()));
-        arguments.add(Integer.toString(comment.countryId()));
-        writers[FileNames.COMMENT_ISLOCATEDIN_PLACE.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-        arguments.add(Long.toString(comment.messageId()));
         arguments.add(Long.toString(comment.author().accountId()));
-        writers[FileNames.COMMENT_HASCREATOR_PERSON.ordinal()].writeEntry(arguments);
+        arguments.add(Integer.toString(comment.countryId()));
+        if (comment.replyOf() == comment.postId()) {
+            arguments.add(Long.toString(comment.postId()));
+            arguments.add(empty);
+        } else {
+            arguments.add(empty);
+            arguments.add(Long.toString(comment.replyOf()));
+        }
+        writers[FileNames.COMMENT.ordinal()].writeEntry(arguments);
         arguments.clear();
 
         for (Integer t : comment.tags()) {
@@ -313,24 +243,10 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
         arguments.add(empty);
         arguments.add(empty);
         arguments.add(Integer.toString(0));
-        writers[FileNames.POST.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-
-        arguments.add(Long.toString(photo.messageId()));
-        arguments.add(Integer.toString(photo.countryId()));
-        writers[FileNames.POST_ISLOCATEDIN_PLACE.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-        arguments.add(Long.toString(photo.messageId()));
         arguments.add(Long.toString(photo.author().accountId()));
-        writers[FileNames.POST_HASCREATOR_PERSON.ordinal()].writeEntry(arguments);
-        arguments.clear();
-
-
         arguments.add(Long.toString(photo.forumId()));
-        arguments.add(Long.toString(photo.messageId()));
-        writers[FileNames.FORUM_CONTAINEROF_POST.ordinal()].writeEntry(arguments);
+        arguments.add(Integer.toString(photo.countryId()));
+        writers[FileNames.POST.ordinal()].writeEntry(arguments);
         arguments.clear();
 
         for (Integer t : photo.tags()) {
@@ -365,5 +281,4 @@ public class CSVPersonActivitySerializer extends PersonActivitySerializer {
     public void reset() {
         // Intentionally left empty
     }
-
 }
